Add bcftools.1 man page extracted from samtools.1
Samtools has provided a unified samtools/bcftools man page since
release 0.1.17.  Now that bcftools is a separate project, move the
bcftools sections of the documentation to here, alongside the code
that they document.

This old documentation pertains to the old samtools/bcftools program,
which differs substantially from the current htslib-based bcftools.
But hopefully it's a useful beginning.