Remove unused recursive Makefile infrastructure
With SUBDIRS=. there is no need for this *-recur infrastructure.
Moreover with bcftools depending on a phony target, it was needlessly
rebuilt every time, which led to "make install" rebuilding it -- which
is a problem if for example you are building and installing on different
machines/distros/architectures.

Also remove versioning complexity needed by htslib's Makefile but
unneeded here as we're not building a library.