Add an unoptimised GCC C99 build
This build complains about plain (non-static) inline functions, the
interpretation of which differs between gcc -std=gnu90 and C99.

Also redo the subprojects as a before_script, so $HTSDIR/$PATH trivia
doesn't appear in the build matrix and we're not overriding the install
script (though that's not otherwise being used anyway).