Use standard "clean" targets; disconnect clean-htslib
Use clean/testclean rather than cleanlocal/cleantest; this is in line
with the targets in the htslib and samtools makefiles and in the spirit
of http://www.gnu.org/prep/standards/html_node/Standard-Targets.html .

Make "clean" is now just a local clean; use "clean-all" to clean htslib
as well, as desired at meetings past and as will be coming soon to
samtools (at last!).

Add "tags" target.