Move test_rename_chrs() output files out of $$opts{tmp}
Evidence of test failure evaporates if it is written to the tmp
directory (which is cleaned when the script ends), so write the files
to $$opts{path} like all the others.  Unlike the others, rename.out.tmp
is created each time in test_rename_chrs() rather than prewritten,
and is usually deleted afterwards (except when renaming tests fail).