Print licensing info in --version and usage
The license under which a bcftools executable may be distributed
depends on whether it uses the GNU Scientific Library (which is
currently equivalent to whether it includes the polysomy command).

Display the license in --version output according to the conventions
at http://www.gnu.org/prep/standards/html_node/_002d_002dversion.html
and also display it in usage output in a GSL-using executable.