Update bcftools.txt
I was pretty sure it would be possible to extract biallelic SNPs with bcftools, but the word biallelic appears nowhere in the documentation. I found the functionality by searching for the string allel. I added an example of what I imagine is a frequent use of bcftools. I'm happy to move this to an example/FAQ page instead to keep the documentation clean/minimal/streamlined. Thanks.