[Makefile] Use user's CPPFLAGS/LDFLAGS/LIBS variables
Use these variables in relevant compilation and linking rules so they can
be overridden by the user in the usual autoconf-style way.  Fixes #222.

Remove DFLAGS and INCLUDES, using the conventional CPPFLAGS instead.
Add EXTRA_CPPFLAGS in their place, for our own various options that
need to stay set even if the user overrides CPPFLAGS.  If we end up
with necessary compiler or linker options or libraries, we'll add
EXTRA_CFLAGS, EXTRA_LDFLAGS, or EXTRA_LIBS accordingly.

Rename *_LDLIBS to *_LIBS; it seems that LDLIBS doesn't particularly
exist outside GNU Make's built-in rules.  Instead LIBS is configure's
conventional variable for the purpose, so is the right thing to use
for a user-visible variable.