plot-vcfstats: plotting of sample names fixed
Plotting of sample names instead of sample ID numbers (flags -s/--sample-names) in plot-vcfstats did not work as the first input to plt.xticks in <prefix>-plot.py was an array of strings instead of an array of integers for all by-sample plots. A simple call of int() around the row[0] argument in the list comprehension fixes that.