Change behavior of `mpileup -d, --depth`
The original -d heuristics assumed one multisample bam and automatically
adjusted the depth up to the total of 8000 pileup reads. However, too big
numbers caused overflow and lowered the maximum depth.

Because we can get any combination of single- and multi-sample bams, the
automatic behavior can be problematic and impossible for the user to
override. Therefore this commit disables the automatic adjustment and
instead prints an informative warning message.