Add configure options for plugins
Allow plugins to be enabled/disabled via configure, and add options to
set the install directory and default plugin search path.

If plugins are enabled:
 * The plugins are built and installed
 * `make test` will test the plugins
 * 'plugin' will appear in the bcftools usage output
 * `bcftools plugin` will work as normal

If plugins are disabled:
 * The plugins are not built
 * `make test` does not try to test the plugins
 * 'plugin' is hidden from the bcftools usage
 * `bcftools plugin` prints a short message to say that plugins are
   disabled, and exits non-zero.

If configure is not used, the default is to enable plugins.

The Makefile is rearranged slightly so that default values and overrides
from config.mk work correctly.  The platform test in the Makefile is
changed so that it can be overridden by configure, so that the
correct result should be used when cross-compiling.