Add --enable-libgsl option to autoconf
Add a configure option to allow linking against the GNU Scientific
Library, enabling the polysomy subcommand.  Includes library checks
for libgsl and cblas.  The cblas library may be called libcblas or
libgslcblas. The default is to look for both of these, but a
--with-cblas option is also included to allow the cblas library name
to be selected manually.

A message about the change of license to GPL is printed out if configure
is run successfully with --enable-libgsl.

Travis in autoconf mode is updated to use --enable-libgsl.