Add configure test for -rdynamic
Test in configure to see if the compiler will accept -rdynamic.
If it does, DYNAMIC_FLAGS in Makefile will be set to -rdynamic,
otherwise it will be unset.  The default if configure is not used
is to set it, as this seems to work best on most of the platforms
we target.