Fix a recently introduced bug,
do not attempt to guess automatically the ploidy, instead allow
the user to provide explicit rules. This solves the problem
of missing genotypes vs ploidy=0 cases.

In order to be able to list predefined rules, the plugin had
to change from the init() to the run() API, which means
the default plugin options are (temporarily?) lost. Easy
to add if required, albeit with different short options
due to conflicting -r,-R,-t,-T.

Fixes #715 and #698.