Prevent dropping AC,AN INFO fields on -i/-e MAF,AF queries
The query command asks htslib to skip the expensive parsing of INFO and FORMAT
fields when it is not needed in the formatting and the filtering expression.
This backfired for fields calculated on the fly, commands like

    bcftools query -e 'MAF[0]<0.05' -f '%POS\n'

would not bee able to calculate MAF because INFO/AC, INFO/AN is not accessible.

This is now fixed, but with a caveat: if the header contains INFO/AC,AN line
but the annotations are not filled in the data lines, the program will not
be able to determine AC, AN from FORMAT/GT. In such case, fill-tags can
be used to update the INFO annotations or the tags can be stripped from the
header using `annotate` or `reheader`.

Resolves #762.