Reduce `csq` memory usage
The haplotype-level consequence calling requires bufferring of VCF lines until
all active transcripts have been processed. This can be a problem when working
with VCFs with tens of thousands samples, the buffered lines can take 15GB of
memory or more.

The commit reduces memory usage significantly (by 90%) when run in the local
calling mode.

This commit also reduces the memory usage in the haplotype-level calling mode,
but here the reduction is not large, approximately ~20%, depending on the data
and the transcript length.  A temporary external swapping solution may be
necessary in future if bigger data sets need to be processed.

The commit also fixes a small memory leak and a bug in FORMAT/BCSQ formatting
at sites with many consequences.