Use double-precision floats when comparing big numbers.
Numeric comparisons in filtering expressions use internally the same code for both integer and floating point
arithmetics. This is a problem when comparing genomic coordinates, the result become imprecise because
single-precision floats can represent integers precisely only up to 16777216. As explained in #871, it is
not always better to use doubles.

This commit improves eded7db and 44192ea by checking if big positive numbers are compared and switch
to double-precision comparison in such case. In the context of VCF, genomic coordinates are the only
but important use case for this exception.