Add new `bcftools call -G, --group-samples` option
This allows to address a long-standing problem with variant calling from low-coverage data: the HWE assumption
is applied to all pooled samples and calls non-reference genotypes even in samples with no alternate alleles.
The only possible way to work around would be to run a single-sample calling or call by population. However,
that's inconvenient and leads to problems with unifying indel alleles.

The new -G option allows to to group samples into populations and apply the HWE assumption within but not
across the groups.