Don't add hard-coded -ldl if configure was used
On OpenBSD (as on macOS) there is no -ldl to link against, so having it
in $(DL_LIBS) causes linking to fail and needs to be avoidable by running
configure. On other platforms, it will be added by configure as needed
and adding it in $(DL_LIBS) causes (harmless) duplication.

But keep it if configure is not run, as most common platforms handled
by this branch of the ifeq/.../else (i.e., Linux) do require it.