Add `bcftools view --with-header` option
This parallels samtools view's --header-only/--no-header/--with-header
trio of options. It is the same as the default, but useful to reset the
effects of --header-only/--no-header.

Also write -h/-H on separate usage lines for clarity, and reformat nearby
usage lines accordingly. (View with diff -w to see the real differences.)