New -C option and `append-missing` logic
The -C option allows to read a long list of options from a file to prevent impossibly long command lines.

The `append-missing` logic allows annotations to be added for each ALT allele in the same order as they appear
in the VCF. Note that this is not bullet proof. In order for this to work:
- the annotation file must have one line per ALT allele
- fields must contain a single value as multiple values are appended as they are
  and would break the correspondence between the alleles and values