[NEWS] Major revamp of `bcftools +trio-dnm`
The original trio-dnm calling model used genotype likelihoods (PLs) as the
input for calling. However, that is flawed because PLs make assumptions which
are unsuitable for de novo calling: PL(RR) can become bigger than PL(RA) even
when the ALT allele is present in the parents. Note that this is true also
for other programs such as DeNovoGear which rely on the same samtools calculation.

The new recommended workflow is

    bcftools mpileup -a AD,QS -f ref.fa -Ou proband.bam father.bam mother.bam |
       bcftools call -mv -Ou |
       bcftools +trio-dnm -p proband,father,mother -Oz -o output.vcf.gz

This commit also implements the DeNovoGear model. The original behavior of trio-dnm
is no longer supported.

For more details see http://samtools.github.io/bcftools/trio-dnm.pdf