Improve mpileup speed by delaying BAQ calculation.
Also a minor refactoring of bcf_call_glfgen for speed.

BAQ is the dominant CPU time for mpileup.  However it's not needed for
the bulk of reads.

We delay BAQ until we observe that a read has an overlap with others
containing an indel.  No indels means no BAQ.

Furthermore, if we do have an indel, but this particular read spans
the indel with more than 50bp either side then also don't realign it.
This is based on the observation that most misalignments occur near
the ends of reads where the optimal pair-wise alignment (read vs ref)
differs from the optimal multiple alignment (read vs all others).

Evaluations so far show this is around a 1.5 to 2-fold speed increase
(depending on tuning parameters), while simultaneously slightly
increasing SNP calling accuracy.

The latter appears to be due to BAQ in non-indel regions sometimes
reducing all reads to QUAL 0 in low complexity areas, blocking SNP
calling from working.  SNPs flanking indels also get blocked for
similar reasons.

This has two variants (see REALN_DIST #if) as an exploration of
methods.

This is enabled with a new -D option.
By default it is as before, and -B still disables completely.

If we like how it works though it's possible we could make this the
default and add a +B for all-BAQ, -B for no-BAQ and have the default
inbetween as auto-select BAQ.

TODO: Consider adding a depth filter as low depth portions aren't
optimal.

Honour MPLP_REDO_BAQ in mpileup