Use a directional MWU test.
The old p(U) calc gave eg <-4 and >4 both as near 0, but for some
calculations we expect a bias in one direction and not the other.  Eg
MQB (mapping quality bias) is skewed as REF matching reads have higher
MQUAL and ALT matching reads.

The U and Z scores are now selectable via #define using the same
function, but U is default for compatibility.