Add an indel-bias argument to mpileup.
This is a multiplicative factor (albeit 1/x) on the BAQ score in the
indel caller.  Given the constant, arbitrary, cap of 111 this is
more or less equivalent to permitting higher scores.  The effect is
to be more permissive in calling, giving greater recall rate at a cost
of slightly reduced precision.

A value of 1 would match the previous computation, but the new default
is 2 (=> score * 1/2).  On a 30x GIAB HG002 chr1 sample filtering a
Q20 this removes 1/3rd of the false negatives, but at a cost of about
1/3rd more FP.  At 60x it's about -40% FN + 20% FP. (Given the
quantity of true calls hugely outweights false, this is a reasonable
trade.)

Also adjusted the default -h parameter from 100 to 500. This gives a
huge reduction in false negative rates (as much as 3 fold), but also
usually has a slightly beneficial decrease in FP.  This parameter is
used to indicate the likely sequencing error rates in homopolymers,
which (for Illumina instruments at least) have greatly reduced since
the original bcftools indel caller was written.  High -h parameters
though does marginally harm SNP calling precision, but not massively.