Add indel BQBZ, MQBZ, RPBZ and SCBZ INFO fields.
These values weren't captured in the indel code so we can't
post-filter on them.  It's hard to know what is ALT and what is REF,
but for now we take the naive approach of a read that doesn't have an
indel is REF and ALT otherwise, irrespective of quality of match.

Of these, MQBZ shows some discrimination (< -5 for 30x and < -9 for
150x), but RPBZ and SCBZ have overlapping distributions (although RPBZ
may just about have a slim margin worthy of filtering at the right
end).  That said for both the dist of the true variants extends much
further negative than the dist for the false ones, so while we cannot
remove FP without removing TP, we could at least use this for
computing a refined QUAL value.

[BQBZ undergoing testing.]