Normalise indel iscore mod by number of reps.
This is length / rep_len.  The theory being the +score is proportional
to the copy number rather than total length, as that is a better
indication of the number of different alignment possibilities.

In practice it seems slightly better discriminating, although the jury
is out whether that's simply due to score being lower on average.