Fix a bug with the placeholder _Q (ZQ) tag.
As it's held in static memory, it needed to put the default '@' symbol
back after nulling it.  This of course makes it not thread safe, but
we don't have any code to run this in a threaded environment.  The
alternative is a non-static buffer and memset the appropriate length
each time.  I haven't tested what overhead that brings.