Fix for REF/ALT indel calculation.
The fixes in commit 3c1205c1 while still here, no longer worked in
this branch due to the swapping of seqQ and baseQ.  The fix however
seemed inappropriate and caused other issues, such as #1446.

The new logic is to check p->indel to decide if the read needs b=0
(REF) setting rather than a naive qual check.  This greatly reduces
the false positive rate of indel calling, especially when deep,
however it also increases GT errors a bit.  I've tried to get some
middle ground where the new seqQ used for REF calls isn't mostly the
old qual, but with a partial mix in of the base qual.  I don't like
heuristics, but it works well on the data sets tested and fixes both
the issue addressed by 3c1205c1 and also #1446, while keeping the FP
reduction.

Beneficial at 30x, 60x and 150x both filtered and unfiltered (huge
win).  At 15x it's a small win for unfiltered and a moderate loss for
filtered, almost entirely due to higher GT error rate.  Possibly this
is filtering rules no longer appropriate, or maybe it's simply not
good on shallower data.  Needs investigation.