Correct the mpileup --config 1.12 parameter set.
I forgot to add -F0.002 in here.  The new default is -F0.05.

Note this also may be a problem for multi-sample calling.  A figure of
5% is too high once you start having lots of files, so I think perhaps
-p should be the default if multiple files are specified. (It's useful
in a single file scenario because it's a pre-filter and cuts out a lot
of the realignment assessment.)  If we do this we'd need the negation,
mayube --combine-sample-mF long option.