Add new --ar, --ambig-reads option
which allows to retain the original behavior (the default) or
increment AD counters proportionally (--ar incAD) or just the
reference counter (--ar incAD0).

Note that the reference counter is not incremented explicitly by the
incAD0 mode, for now we are assuming that it is always the case that
the competing indel allele is the reference allele. When this is not
the case, the correct counter will be incremented, only the 'incAD0'
is somewhat misleading in such (rare) cases.