Exit gracefully instead of segfaulting
When `bcftools annotate --set-id` and `--remove` are combined,
`--set-id` cannot use tags deleted by `--remove`. This is now
detected and the program exists with an informative error
message.

A possibile enhancement would be to make the order of operations
customizeable, for example by the order of the options on the command
line.

Resolves #1540