Resolve a "fixme" case in `call -C alleles`
When constraining sites w/o the unseen allele <*> to a new allele,
a "fixme" error was triggered. This commit uses the heuristic of
switching to the likelihoods of last alternate allele in the list,
which should be the least likely. Although not mathematically correct,
this allows to proceed despite such sites.

Resolves #1557