New `--ligate-force` and `--ligate-warn` options
for finer control of the `-l, --ligate` behavior in
imperfect overlaps. The new default is to throw an error
when sites present in one chunk but absent in the other
are encountered. To drop such sites and proceed, use
the new `--ligate-warn` option (previously this was the
default). To keep such sites, use the new `--ligate-force`
option.

Resolves #1567