Add new --indel-size option
This allows to increase the maximum indel size considered, large
deletions in long read data are otherwise lost.

In this initial commit this is optional and not considered by any
of the presets. Once this is benchmarked and performance cost is known,
it might be good to increase the default value with `-X ont` and
`-X pacbio-ccs`.

Resolves #1569