Change the MWU test so MWU 1 or MWU-Z 0 is returned for variance=0.
When we only have the one data value, eg 50 alt and 50 ref all with
MQual 60, this gives a variance of 0.  To avoid the division by zero
later, this was simply returned as HUGE_NUM and produces no output
value to filter on.

Instead we now output a probability to explicitly state no bias.

This has no impact when filtering out (view -e), but does when
filtering-in with view -i.