Further speed up vcf merge.
maux_reset() spams bcf_hdr_name2id, looking up the same chromosomes
over and over again.  It does this because they may be in a different
order for each of the files being merged, but we can at least cache
the last rid value per file and reuse when the chr doesn't change.

This is another 9% reduction in time over the previous commit.