Introduce a new option --no-indelQ-tweaks for more sensitive indel calling
The option switches off an existing heuristics which reduces indelQ
and, in the extreme case, has the power to discard an indel
entirely when the reference alignment has low score.

This is a problem for long reads, so newly `--no-indelQ-tweaks`
is added to 'ont' and 'pacbio-ccs' presets.

Tentative fix to #1459