Tweaks to try and improve on insertion calling.
The define INS_PLUS_BASE is enabled, which is how the old code worked
so no change.

This includes the next base matching call in the last base of the
insertion, so it's *either* cons_base *or* cons_ins.

However we instead have an easier to understand alternative which is
an easier to understand with cons_ins being the inserted bases and
cons_base always being used for the matching bases.  (This corrects
some defects for reads than end on an insertion with no flanking M,
although that is rare.)

Unfortunately, for reasons unknown, it is very slightly higher on
FP/GT. (Hence we still use the old method)