Trivial fix to flush VCF/BCF before erroring on index writes.
Failing to write an index is an error, but we're likely to get bug
reports if we write 99% of a VCF or BCF and then error without
flushing the remainder.  People will just run the index command as a
separate step and then be confused as to why the get broken output.