Give control over creation of vectors with mixed known and missing values
When input files have different alternate alleles, vector fields pertaining to
unobserved alleles are set to missing by default. This creates vectors with
mixed known and unknown values and some programs refuse to work with such.

This commit changes the default behavior of --gvcf merging:

- whenever the uknown allele is present (<*> or <NON_REF>) its values are
  used instead of the missing value
- when it is not present, the default rule `-M PL:max,AD:0` is used

A new `-M, --missing-rules` option is added which allows to override the
default `-M PL:max,AD:0` rule implied by `--gvcf`. Note that the use of
the unknown allele (<*> or <NON_REF>) given explicitly cannot be overriden,
which I believe is the correct behavior.

Resolves #1888.