Identical rows must be returned when `-s` is applied regardless of `-f` containing the `-a` VEP tag itself or not.
The problem was caused by using unparsed INFO/CSQ in one instance and
parsed INFO/CSQ restricted to the desired `-s` transcipt in the other.

For example, this command

    bcftools +split-vep test/split-vep.filter.vcf -s worst -i'CSQ~"nonsense"' -f '%POS %Consequence %CSQ'

queried only the INFO/CSQ part related to the worst transcript (the desired behavior),
while this command

    bcftools +split-vep test/split-vep.filter.vcf -s worst -i'CSQ~"nonsense"' -f '%POS %Consequence'

queried the whole INFO/CSQ string.