Allow combining -m and -a with --old-rec-tag
If --old-rec-tag was requested with atomization, the output header was
updated with the requested tag, however the input header is used when
splitting the atomized lines with `-m -`. This lead to a segfault because
the new tag added by atomization was not present in the input header.

This fixes the problem by using the output header for both input and output.
As long as future code only adds and never removes or changes ids of existing
tags, this should be a robust solution. (Expect troubles if this assumption
is broken, we don't have good API to support such situations.)

Fixes #2020