Add a new indel caller.
This is a combination of PR #1679 from 2022 coupled with more recent
changes to replace BAQ with edlib for evaluating the alignment of
reads against candidate alleles.  Due to the complexity of rebasing
many commits with many conflicts, these have been squashed together
for simplicity.  Please consult the original commit log in the PR for
more detailed history.

PR 1679 is a major restructuring of the indel caller that produces
proper consensus sequences.  TODO: paste collected commit messages
here.

The newer edlib changes are designed to dramtically increase the speed
of indel calling on long read data by use of a modern algorithm.  Note
edlib's alignments are very crude, with just +1 for all differences
including substitutions and each indel base.  However if we've
computed our candidate allele consensus sequences correctly then the
aligning against the real allele should be a minimal score regardless
of scoring regime, so the cost of an HMM or even affine weights do not
help much.

The lack of quality values is compensated for by detection of minimum
quality within STRs.