- Enable the "edlib mode" for the more recent mpileup -X profiles.
It's faster and more accurate than the old mode.  The speed is due to
BAQ vs Edlib, but the accuracy comes from the use of allele consensus
generation and better heuristics.

- Cull lots of ifdefed out code, plus reset indel-bias back.

Earlier experiments had increased this score by doing indel_bias/10
when it's used (it's 1/score for effect).  The difference isn't great,
and it's marginally better FN/FP ratio with it in the original usage.
We can still manually change this score to get better sensitivity if
desired, but it's not dialed up so high by default.

- Remove an indel caller hang.

Our consensus can sometimes fail and give us a short alignment that
doesn't span pos.  We were checking the wrong end coord.