Fix the indelQ assignment for multi-allelic indel sites.
When we have eg REF C and ALT CT,CTT then we're saying it's +1 or +2
Ts.  We align all reads against the 3 candidate alleles and score
them, sorting these scores to find the best so we can allocate reads
to the alleles.

The score for those allele assignments was REF minus best-indel for
REF assignments, and best-indel minus REF for indel assignments.  This
change turns the latter into best-indel vs next-best-allele,
regardless of whether that next best is REF or not.

Consider the case of +1 score 30, +2 score 30, +0 (REF) score 60.
Previously we'd have recorded the relative quality of 60-30, but now
we record 30-30.  The consequence of this is reads that align equally
well against +1 and +2 get zero confidence in their correct allele
assignment.

This considerably reduces the chance of recording GT 1/2 for variable
homopolymers caused purely by the vagarities of sorting a bunch of
equal numbers.  (The equal scores often arrive due to reads that don't
span the homopolymer or STR.)  We may miss some true variants as the
multi-allelic possibilities all cancel each other out, but typically
that means there are very few reads spanning the site and this data
may well be false-positive caused by sequencing artifacts.  So overall
it's a considerable benefit to accuracy.

A consequence of this is also that AD values now more accurately
reflect evidence, rather than incorrectly distributing the
uninformative reads to alleles.

However, AD will now also be lower, and similarly PL.  So a global
boost to indelQ helps recover some of this lost sensitivity.  This
commit also boosts indelQ by the --indel-bias param to make this more
impactful. (Previously it only boosted seqQ)