- Rename mpileup --edlib to --indels-cns
The main benefit of the new indel caller is the use of alignment
against diploid consensus generation instead of simply patching the
reference with candidate indel types.

This greatly reduces false positives and incorrect allele alignment
(leading to wrong genotype calls).  This was the earlier PR #1679, but
has since acquired edlib as an alternative to BAQ for the indel
alignment.  However this is primarily a speed benefit (with some
minor removal of false-negatives due to quality smashing), and
isn't the main thing users should think about when choosing an indel
caller.

Also tidied up the usage statement and added an explicit "-X list" to
print the profile parameters.

- Add extra debugging defines.

GLF_DEBUG reports GLF calculation in bam2bcf.c.

ALIGN_DEBUG uses edlib EDLIB_TASK_PATH to report sequence alignment.
NB to use this you need to link against edlib itself rather than the
cutdown version in this repository.

Also fix the edlib heuristics used in bcf_call_glfgen.  We don't want
to change the call (b=5) as this affects AD.  Instead we change the
quality so poor calls get filtered by QUAL rather than simply being
removed.

- Tweak edlib tuning for SeqQ/qual.
Add quality value assessment into soft-clip recovery.
Use /500 instead of /111 in indelQ assignment, and skew indel-bias
accordingly.  This gives better separation of FP/GT/FN generally.

- Added --seqq-offset parameter so we can use it in tunables per
profile.  This is used as a limit on the seqQ reduction in the
"VAL-5*MIN(20,depth)" formula, used for favouring data over seqQ
scores when depth is sufficient.  Experimentation showed no single
value that worked for all platforms, but the default is in the
middle.

- Tidy up to cull ifdefed and commented out code.

- Add test for indels-cns.

It's minimal, but the whole indel calling has minimal testing.  I
think we have under 10 indels in total with develop (all short read
and mostly duplications of each other), and no testing of indels-2.0.

This tests 4 indels with indels-cns.

- Added documentation for the new --indels-2.0 options

- Cull more unused parts of edlib.c.

This avoids clang warnings (which become errors with -Werror).
We're only including the bits we need here for mpileup.  If you want
the whole thing, link against the upstream source library instead.