Don't change indel qual when indelQ == 0.
For an alignment that doesn't have an indel but is aligned against
reads that do have an indel, the indel quality comes from the BAM
quality.  However we already have indelQ assigned, so this avoids
changing to BAM qual if indelQ is zero as that is a special case for a
read aligning to multiple indel "types" (lengths) with equal score.

This avoids excess AD numbers for poorly chosen alignments.

Fixes #2113