Tweak the condition for calling a 2nd consensus sequence in a deletion.
CONS_CUTOFF_DEL and CONS_CUTOFF_INC are now 35% instead of 40% (also
DEL is new as it previously shared the score with SNPs).

Plus when computing the ratio we no longer apply this as a percentage
to the entire depth, but as a percentage of the alignments overlapping
any STRs spanning this region (so meaningful alignments) that aren't
already allocated to the "type" being diagnosed.

Both of these make it slightly more likely to emit a second consensus
sequence potentially containing indel variations.  This fixes the
specific problem raised in #2117, but like all heuristics it benefits
some cases and harms others.  We will always have some cases that fall
just either side of the thresholds and look odd.

Overall it seems a slim benefit and the work to track the number of
meaningful spanning alignments may have potential value in subsequent
updates.

Fixes #2117