Fix REF indel calls for multi-sample calling.
This applies where there are zero observed indels in a sample (but
they are in other samples).  It's odd how this appears to be a pretty
rare change, as we'd expect a significant change to FP, but it's
tiny.  Genotype assignment error changes a lot, but only in one of the
3 samples I tested.

On HG002, I see the following differences to calling rates.

Previous     All     QUAL>=100
InDel TP   11823 /   11798
InDel FP    5374 /    4898
InDel GT     296 /     291
InDel FN     115 /     140

New          All     QUAL>=100
InDel TP   11822 /   11795
InDel FP    5313 /    4805
InDel GT      80 /      74
InDel FN     116 /     143

This was HG002 called in conjunction with HG003 and HG004, but not as
a trio (so no pedegree supplied).  Oddly despite that HG002 is much
more accurate than HG003 and HG004, with GT assignment error rates an
order of magnitude higher.  This PR makes them a bit higher still
(maybe another 20%).  I cannot explain either of these, but perhaps
it's simply down to the accuracy of the truth set as HG002 is by far
the most widely curated of the three.  Either that or my analysis has
a flaw somewhere.

Fixes #2130.