Make --write-index have an optional =FMT option.
Using --write-index defaults to --write-index=csi as before, but we
can now do --write-index=tbi to get TBI indices instead.  This takes
precedence over auto-detection based on filename suffix when using the
filename##idx##indexname nomenclature.

The exception to defaulting to CSI is bcftools isec, which defaults to
TBI.  This disparity was there before and this PR doesn't change that
behaviour.

Fixes #2008